update instructions to be platform agnostic
We should move our focus to vanilla kubernetes and provide links to methods to
instaling and using the various platforms (minikube, BKS, ICP). Further howtos
can be done to focus on utilizing a specific platform for it.

Also remove the replication set instructions, we shouldn't confuse the audience
by providing two methods to install when only one is really practical.

Also fixed the LICENSE file.